Avoid rate aberations in market analysis
Current implementation always take the lowest rate in offers.

There is often some "aberations" data before the pack of offers with a
similar rate.

This commit removes the offers that are bellow the 5th percentile of the
30 first offers.

Consider the following offers:

- 0.00435
- 0.00597
- 0.00597
- 0.00597
- 0.00598
- 0.00599
- 0.00599
- 0.00599
...

The goal is to remove `0.00435`.

By not doing so, it takes the whole Poloniex rates lower.